List empty repositories (#33)
# Changes

## Primary change

Add a query to list repositories. It only returns an empty listing,
because there is not yet a way to register repositories.

## Supporting changes

Re-organized everything by domain-driven design, once I encountered a
circular dependencies error. Each package's role is documented in
`<package>/doc.go`.

## Future work

Add a command to register repositories, so it will be necessary to write
JSON to files and read it back when listing repositories.

## Version

Same.